Refactor: Reindex -> Index reindex -> index REINDEX -> INDEX
