+++ molgenis-data-elasticsearch/src/main/java/org/molgenis/data/elasticsearch/index/job/IndexJobExecutionFactory.java
@@ -1,0 +1,17 @@
+package org.molgenis.data.elasticsearch.index.job;
+
+import org.molgenis.data.AbstractSystemEntityFactory;
+import org.molgenis.data.elasticsearch.index.meta.IndexJobExecutionMeta;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+@Component
+public class IndexJobExecutionFactory
+		extends AbstractSystemEntityFactory<IndexJobExecution, IndexJobExecutionMeta, String>
+{
+	@Autowired
+	IndexJobExecutionFactory(IndexJobExecutionMeta indexJobExecutionMeta)
+	{
+		super(IndexJobExecution.class, indexJobExecutionMeta);
+	}
+}
