+++ molgenis-data/src/main/java/org/molgenis/data/index/meta/IndexActionGroupFactory.java
@@ -1,0 +1,16 @@
+package org.molgenis.data.index.meta;
+
+import org.molgenis.data.AbstractSystemEntityFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+@Component
+public class IndexActionGroupFactory
+		extends AbstractSystemEntityFactory<IndexActionGroup, IndexActionGroupMetaData, String>
+{
+	@Autowired
+	IndexActionGroupFactory(IndexActionGroupMetaData indexActionGroupMetaData)
+	{
+		super(IndexActionGroup.class, indexActionGroupMetaData);
+	}
+}
