--- molgenis-data/src/main/java/org/molgenis/data/reindex/meta/IndexPackage.java
@@ -1,33 +1,0 @@
-package org.molgenis.data.reindex.meta;
-
-import org.molgenis.data.meta.SystemPackage;
-import org.molgenis.data.meta.model.PackageMetaData;
-import org.molgenis.data.system.model.RootSystemPackage;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
-import static java.util.Objects.requireNonNull;
-import static org.molgenis.data.system.model.RootSystemPackage.PACKAGE_SYSTEM;
-
-@Component
-public class IndexPackage extends SystemPackage
-{
-	public static final String SIMPLE_NAME = "idx";
-	public static final String PACKAGE_INDEX = PACKAGE_SYSTEM + PACKAGE_SEPARATOR + SIMPLE_NAME;
-
-	private final RootSystemPackage rootSystemPackage;
-
-	@Autowired
-	public IndexPackage(PackageMetaData packageMetaData, RootSystemPackage rootSystemPackage)
-	{
-		super(SIMPLE_NAME, packageMetaData);
-		this.rootSystemPackage = requireNonNull(rootSystemPackage);
-	}
-
-	@Override
-	protected void init()
-	{
-		setLabel("Index");
-		setParent(rootSystemPackage);
-	}
-}
