--- molgenis-data/src/main/java/org/molgenis/data/reindex/meta/ReindexActionFactory.java
@@ -1,15 +1,0 @@
-package org.molgenis.data.reindex.meta;
-
-import org.molgenis.data.AbstractSystemEntityFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
-@Component
-public class ReindexActionFactory extends AbstractSystemEntityFactory<ReindexAction, ReindexActionMetaData, String>
-{
-	@Autowired
-	ReindexActionFactory(ReindexActionMetaData reindexActionMetaData)
-	{
-		super(ReindexAction.class, reindexActionMetaData);
-	}
-}
