--- molgenis-data/src/main/java/org/molgenis/data/reindex/meta/ReindexActionGroupFactory.java
@@ -1,16 +1,0 @@
-package org.molgenis.data.reindex.meta;
-
-import org.molgenis.data.AbstractSystemEntityFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
-@Component
-public class ReindexActionGroupFactory
-		extends AbstractSystemEntityFactory<ReindexActionGroup, ReindexActionGroupMetaData, String>
-{
-	@Autowired
-	ReindexActionGroupFactory(ReindexActionGroupMetaData reindexActionGroupMetaData)
-	{
-		super(ReindexActionGroup.class, reindexActionGroupMetaData);
-	}
-}
