+++ molgenis-gavin/src/main/java/org/molgenis/gavin/job/meta/GavinJobExecutionMetaData.java
@@ -3,14 +3,14 @@ package org.molgenis.gavin.job.meta;
 import org.molgenis.data.jobs.model.JobExecutionMetaData;
 import org.molgenis.data.meta.SystemEntityMetaData;
 import org.molgenis.data.meta.model.AttributeMetaDataFactory;
-import org.molgenis.data.reindex.meta.IndexPackage;
+import org.molgenis.data.index.meta.IndexPackage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import static java.util.Objects.requireNonNull;
 import static org.molgenis.MolgenisFieldTypes.AttributeType.STRING;
 import static org.molgenis.data.meta.model.Package.PACKAGE_SEPARATOR;
-import static org.molgenis.data.reindex.meta.IndexPackage.PACKAGE_INDEX;
+import static org.molgenis.data.index.meta.IndexPackage.PACKAGE_INDEX;
 
 @Component
 public class GavinJobExecutionMetaData extends SystemEntityMetaData
